complete refactor for price and style
