+++ client/src/components/ProductInfo.jsx
@@ -1,114 +1,77 @@
 /* eslint-disable import/extensions */
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import axios from 'axios';
+import { useSelector, useDispatch } from 'react-redux';
+// import { useGetProductInfoQuery, useGetStylesQuery } from '../reducers/Example-Api-Slice';
+import { updateProductInfo } from '../reducers/Example-Reducer';
+import { updateStyles } from '../reducers/Style-Reducer';
+
 import Price from './Price.jsx';
 import StyleSelector from './StyleSelector.jsx';
 
-class ProductInfo extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      allProducts: [],
-      selectedProduct: {},
-      allStyles: [],
-      styleIndex: 0,
-      selectedStyle: {},
-    };
-    this.getAllProducts = this.getAllProducts.bind(this);
-    this.getAllStyles = this.getAllStyles.bind(this);
-    this.updateProduct = this.updateProduct.bind(this);
-  }
+const ProductInfo = () => {
+  // const [allStyles, getStyles] = useState([]);
+  const productId = useSelector((state) => state.product.id);
+  const product = useSelector((state) => state.product.productInfo);
+  const allStyles = useSelector((state) => state.style.allStyles);
 
-  // renders the info of the specific product on load
-  componentDidMount() {
-    const { productId } = this.props;
-    this.updateProduct(productId);
-  }
+  const dispatch = useDispatch();
 
-  // componentDidUpdate(prevState) {
-  //   const { allStyles, styleIndex } = this.state;
-  //   if (prevState.allStyles !== allStyles) {
-  //     this.state.selectedStyle = allStyles[styleIndex];
-  //   }
-  // }
-
-  getAllProducts() {
-    const { allProducts, productIndex } = this.state;
-    axios.get('/products')
-      .then((response) => {
-        console.log('get all Products');
-        this.setState({
-          allProducts: response.data,
-          selectedProduct: response.data[productIndex],
-        });
-      })
-      .then(() => {
-        console.log(allProducts.length);
-        this.getAllStyles(allProducts[0].id);
-      })
-      .catch((error) => console.log(error));
-  }
-
-  getAllStyles(productId) {
-    const { styleIndex } = this.state;
-    // refactor to pull from api
+  const getAllStyles = () => {
     axios.get(`/products/${productId}/styles`)
       .then((response) => {
-        console.log(response.data.results);
-        this.setState({
-          allStyles: response.data.results,
-          selectedStyle: response.data.results[styleIndex],
-        });
+        // console.log(response.data.results);
+        dispatch(updateStyles(response.data.results));
+        // getStyles(response.data.results);
+        // this.setState({
+        //   allStyles: response.data.results,
+        //   selectedStyle: response.data.results[styleIndex],
+        // });
       })
       .catch((error) => console.log(error));
-  }
+  };
 
-  updateProduct(productId) {
+  const updateProduct = () => {
     axios.get(`/products/${productId}`)
       .then((response) => {
-        this.setState({
-          selectedProduct: response.data,
-        });
+        dispatch(updateProductInfo(response.data));
+        // this.setState({
+        //   selectedProduct: response.data,
+        // });
       })
-      .then(this.getAllStyles(productId))
+      .then(getAllStyles(productId))
       .catch((error) => console.log(error));
-  }
+  };
 
-  updateStyle(styleId) {
-    const { allStyles } = this.state;
-    const newStyle = allStyles.find((style) => style.id === styleId);
-    this.setState({
-      selectedStyle: newStyle,
-    });
-  }
+  useEffect(() => {
+    // This will be invoked only once.
+    updateProduct();
+  }, []);
 
-  render() {
-    const { selectedProduct, selectedStyle, allStyles } = this.state;
-    if (!selectedProduct) {
-      return <span data-testid="loading">Loading...</span>;
-    }
-    return (
-      <>
-        <div data-testid="resolved" className="overview">
-          <span data-testid="ratings" className="ratings">Star Ratings Placeholder</span>
-          <br />
-          <span data-testid="show-category" className="category">{selectedProduct.category}</span>
-          <br />
-          <span data-testid="show-name" className="product-name">{selectedProduct.name}</span>
-          <br />
-          <span data-testid="show-price" className="price"><Price selectedStyle={selectedStyle} /></span>
-          <br />
-          <span data-testid="show-description" className="description">{selectedProduct.description}</span>
-          <br />
-          <span data-testid="social-media" className="social-media">Social Media Placeholder</span>
-          <br />
-          <div data-testid="style-selector" className="style-selector">
-            <StyleSelector allStyles={allStyles} updateStyle={this.updateStyle} />
-          </div>
+  if (allStyles.length === 0) {
+    return <span data-testid="loading">Loading...</span>;
+  }
+  return (
+    <>
+      <div data-testid="resolved" className="overview">
+        <span data-testid="ratings" className="ratings">Star Ratings Placeholder</span>
+        <br />
+        <span data-testid="show-category" className="category">{product.category}</span>
+        <br />
+        <span data-testid="show-name" className="product-name">{product.name}</span>
+        <br />
+        <span data-testid="show-price" className="price"><Price /></span>
+        <br />
+        <span data-testid="show-description" className="description">{product.description}</span>
+        <br />
+        <span data-testid="social-media" className="social-media">Social Media Placeholder</span>
+        <br />
+        <div data-testid="style-selector" className="style-selector">
+          <StyleSelector allStyles={allStyles} />
         </div>
-      </>
-    );
-  }
-}
+      </div>
+    </>
+  );
+};
 
 export default ProductInfo;
